rename view to dbview
